fix: extracted view constants to separate file to prevent compile errors when using constants in backend
