+++ src/components/SideBar/SideBar.tsx
@@ -5,7 +5,7 @@ import { useRecoilState } from 'recoil'
 import addClassString from '../../../utilities/addClassString'
 import LightHouseLogo from '../../assets/images/lightHouse.svg'
 import LightHouseFullLogo from '../../assets/images/lightHouseFull.svg'
-import { PRIMARY_VIEWS, SECONDARY_VIEWS } from '../../constants/constants'
+import { PRIMARY_VIEWS, SECONDARY_VIEWS } from '../../constants/views'
 import useClickOutside from '../../hooks/useClickOutside'
 import useMediaQuery from '../../hooks/useMediaQuery'
 import useUiMode from '../../hooks/useUiMode'
